+++ NEWS.md
@@ -9,6 +9,7 @@
 
 ## Bug fixes
 
+- The `dunn_test()` documentation is updated to describe the discrepancy between the default behaviour of the `rstatix::dunn_test()` comapred to other packages (`dunn.test` and `jamovi`). The default of the rstatix::dunn_test() function is to perform a two-sided Dunn test like the well known commercial softwares, such as SPSS and GraphPad. This is not the case for some other R packages (dunn.test and jamovi), where the default is to perform one-sided test (#50). 
 - Now, the function `get_summary_stats()` handles the user defined probabilities for grouped data (#78)